updated the raw output data for greatest mod calls

+++ css 342/assig3/RCPeters_GCD.h
@@ -8,25 +8,25 @@
 
 class GCD {
 private:
-    unsigned int I,A,B,R,gcd,modCalls;
+    unsigned long long int I,A,B,R,gcd,modCalls;
 
-    unsigned int findGCD();
+    unsigned long long int findGCD();
 
 
 public:
-    GCD(const unsigned int &i):A(0),B(0),R(0),gcd(0),modCalls(0),I(i){};
+    GCD(const unsigned long long int &i):A(0),B(0),R(0),gcd(0),modCalls(0),I(i){};
     GCD(const GCD &copy);
-    GCD(const unsigned int &a, const unsigned int &b,const unsigned int &i);
+    GCD(const unsigned long long int &a, const unsigned long long int &b,const unsigned long long int &i);
 
-    unsigned int getGCD()const;
+    unsigned long long int getGCD()const;
 
-    unsigned int getModCalls()const;
+    unsigned long long int getModCalls()const;
 
-    unsigned int getA()const;
+    unsigned long long int getA()const;
 
-    unsigned int getB()const;
+    unsigned long long int getB()const;
 
-    unsigned int getI()const;
+    unsigned long long int getI()const;
     friend std::ostream& operator<<(std::ostream &os, const GCD &g);
     bool operator>(const GCD &rhs)const;
     bool operator>=(const GCD &rhs)const;
